+++ packages/components/src/DateTimeInput.tsx
@@ -3,6 +3,7 @@ import classNames from 'classnames';
 import Log from '@deephaven/log';
 import MaskedInput, { SelectionSegment } from './MaskedInput';
 import { getNextSegmentValue } from './DateInputUtils';
+import { addSeparators } from './DateTimeInputUtils';
 
 const log = Log.module('DateTimeInput');
 
@@ -33,13 +34,6 @@ function fixIncompleteValue(value: string): string {
       .replace(/\u2007/g, '0')}${DEFAULT_VALUE_STRING.substring(value.length)}`;
   }
   return value;
-}
-
-function addSeparators(value: string): string {
-  const dateTimeMillis = value.substring(0, 23);
-  const micros = value.substring(23, 26);
-  const nanos = value.substring(26);
-  return [dateTimeMillis, micros, nanos].filter(v => v !== '').join('\u200B');
 }
 
 const removeSeparators = (value: string) => value.replace(/\u200B/g, '');
