+++ packages/components/src/MaskedInput.tsx
@@ -2,7 +2,11 @@ import React, { useMemo, useEffect, useCallback } from 'react';
 import classNames from 'classnames';
 import Log from '@deephaven/log';
 import { useForwardedRef } from '@deephaven/react-hooks';
-import { DEFAULT_GET_PREFERRED_REPLACEMENT_STRING } from './MaskedInputUtils';
+import {
+  DEFAULT_GET_PREFERRED_REPLACEMENT_STRING,
+  fillToLength,
+  trimTrailingMask,
+} from './MaskedInputUtils';
 import './MaskedInput.scss';
 
 const log = Log.module('MaskedInput');
@@ -18,41 +22,6 @@ const SELECTION_DIRECTION = {
  * https://www.fileformat.info/info/unicode/char/2007/index.htm
  */
 const FIXED_WIDTH_SPACE = '\u2007';
-
-/**
- * Fill the string on the right side with the example value to the given length
- * @param checkValue Initial string to pad
- * @param exampleValue Example value
- * @param length Target length
- * @returns String padded with the given example value
- */
-function fillToLength(
-  checkValue: string,
-  exampleValue: string,
-  length: number
-): string {
-  return checkValue.length < length
-    ? `${checkValue}${exampleValue.substring(checkValue.length, length)}`
-    : checkValue;
-}
-
-/**
- * Trim all characters matching the empty mask on the right side of the given value
- * @param value String to trim
- * @param emptyMask Empty mask
- * @returns Trimmed string
- */
-function trimTrailingMask(value: string, emptyMask: string): string {
-  let { length } = value;
-  for (let i = value.length - 1; i >= 0; i -= 1) {
-    if (emptyMask[i] === value[i]) {
-      length = i;
-    } else {
-      break;
-    }
-  }
-  return value.substring(0, length);
-}
 
 export type SelectionSegment = {
   selectionStart: number;
